requirements.txt
alembic==0.9.7
amqp==2.2.2
aniso8601==2.0.0
bcrypt==3.1.4
billiard==3.5.0.3
celery==4.1.0
cffi==1.11.4
click==6.7
cryptography==1.7.2
cssmin==0.2.0
Faker==0.8.10
Flask==0.12.2
Flask-Admin==1.5.0
Flask-Assets==0.12
Flask-Bcrypt==0.7.1
Flask-Cache==0.13.1
Flask-Celery-Helper==1.1.0
Flask-Cors==3.0.3
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-RESTful==0.3.6
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.6
itsdangerous==0.24
Jinja2==2.10
kombu==4.1.0
Mako==1.0.7
MarkupSafe==1.0
peewee==3.0.18
pyasn1==0.4.2
pycparser==2.18
PyJWT==1.5.3
PyMySQL==0.8.0
python-dateutil==2.6.1
python-editor==1.0.3
pytz==2017.3
rabbitmq==0.2.0
six==1.11.0
SQLAlchemy==1.2.2
text-unidecode==1.1
vine==1.1.4
webassets==0.12.1
Werkzeug==0.14.1
WTForms==2.1
